CIWEMB-15: Combine duplicate hook function
